// src/orval/type.schemas.ts
export type GetApiAppointmentParams = {
PageSize?: number;
PageNumber?: number;
AppointmentStatus?: AppointmentStatus;
AppointmentType?: AppointmentType;
};
// ...
export interface UserDto {
  /** @nullable */
  companyName?: string | null;
  createdAt?: string;
  /** @nullable */
  email?: string | null;
  /** @nullable */
  firstName?: string | null;
  id?: string;
  /** @nullable */
  lastName?: string | null;
  /** @nullable */
  phoneNumber?: string | null;
  /** @nullable */
  picFullName?: string | null;
  /** @nullable */
  profilePictureUrl?: string | null;
  roleName?: RoleName;
  /** @nullable */
  roles?: string[] | null;
  updatedAt?: string;
  /** @nullable */
  username?: string | null;
}
// ...
export type StatesAndTerritoriesOfMalaysia = typeof StatesAndTerritoriesOfMalaysia[keyof typeof StatesAndTerritoriesOfMalaysia];



export const StatesAndTerritoriesOfMalaysia = {
  Default: 'Default',
  Johor: 'Johor',
  Kedah: 'Kedah',
  Kelantan: 'Kelantan',
  Malacca: 'Malacca',
  NegeriSembilan: 'NegeriSembilan',
  Pahang: 'Pahang',
  Penang: 'Penang',
  Perak: 'Perak',
  Perlis: 'Perlis',
  Sabah: 'Sabah',
  Sarawak: 'Sarawak',
  Selangor: 'Selangor',
  Terengganu: 'Terengganu',
  KualaLumpur: 'KualaLumpur',
  Labuan: 'Labuan',
  Putrajaya: 'Putrajaya',
} as const;

export interface SellRentRequestDetailDto {
  /** @nullable */
  hourMeter?: string | null;
  id?: string;
  isRent?: boolean;
  location?: StatesAndTerritoriesOfMalaysia;
  /** @nullable */
  machineMake?: string | null;
  /** @nullable */
  machineType?: string | null;
  /** @nullable */
  modelNo?: string | null;
  regCard?: boolean;
  requestId?: string;
}

export interface SasUploadMediaDto {
  /** @nullable */
  fileUrlPrefix?: string | null;
  /** @nullable */
  sasUrlForUploadMedia?: string | null;
}

export type RoleName = typeof RoleName[keyof typeof RoleName];



export const RoleName = {
  Default: 'Default',
  User: 'User',
  Dealer: 'Dealer',
  Distributor: 'Distributor',
  Admin: 'Admin',
  BusinessAffiliate: 'BusinessAffiliate',
  Collaborator: 'Collaborator',
} as const;

export interface ResponseDto {
  /** @nullable */
  errMessage?: string | null;
  isSuccess?: boolean;
}

export interface ResetPasswordRequestDto {
  /** @nullable */
  email?: string | null;
  /** @nullable */
  password?: string | null;
  /** @nullable */
  token?: string | null;
}

export type RequestType = typeof RequestType[keyof typeof RequestType];



export const RequestType = {
  Default: 'Default',
  Buy: 'Buy',
  Sell: 'Sell',
  Rent: 'Rent',
  RentOut: 'RentOut',
  Insurance: 'Insurance',
  Dealer: 'Dealer',
  Collaborator: 'Collaborator',
  Advertisement: 'Advertisement',
  Enquiry: 'Enquiry',
} as const;

export type RequestStatus = typeof RequestStatus[keyof typeof RequestStatus];



export const RequestStatus = {
  Default: 'Default',
  Pending: 'Pending',
  Appointment: 'Appointment',
  PendingPayment: 'PendingPayment',
  PendingDelivery: 'PendingDelivery',
  Completed: 'Completed',
  Incomplete: 'Incomplete',
} as const;

export interface RequestSummaryDto {
  buyRentRequestDetail?: BuyRentRequestDetailDto;
  createdAt?: string;
  id?: string;
  /** @nullable */
  requestorName?: string | null;
  requestStatus?: RequestStatus;
  requestType?: RequestType;
  sellRentRequestDetail?: SellRentRequestDetailDto;
}
// ...
export interface BuyRentRequestDetailDto {
  id?: string;
  interestType?: InterestType;
  isRent?: boolean;
  listing?: AdminListingDto;
  listingId?: string;
  requestId?: string;
}

export interface BusinessAffiliateSummaryDto {
  /** @nullable */
  email?: string | null;
  /** @nullable */
  firstName?: string | null;
  id?: string;
  /** @nullable */
  lastName?: string | null;
  /** @nullable */
  phoneNumber?: string | null;
  /** @nullable */
  profileImageUrl?: string | null;
  /** @nullable */
  username?: string | null;
}
// ...
export type AppointmentType = typeof AppointmentType[keyof typeof AppointmentType];



export const AppointmentType = {
  Default: 'Default',
  ViewFullInspectionReport: 'ViewFullInspectionReport',
  SignBuyerAgreement: 'SignBuyerAgreement',
  MachineInspection: 'MachineInspection',
  SignSellerAgreement: 'SignSellerAgreement',
  SignRentingAgreement: 'SignRentingAgreement',
  SignRentingOutAgreement: 'SignRentingOutAgreement',
  MachineDelivery: 'MachineDelivery',
} as const;

export type AppointmentStatus = typeof AppointmentStatus[keyof typeof AppointmentStatus];



export const AppointmentStatus = {
  Default: 'Default',
  Scheduled: 'Scheduled',
  Completed: 'Completed',
  Cancelled: 'Cancelled',
} as const;

export interface AppointmentDto {
  address?: AddressDto;
  appointmentDateTime?: string;
  appointmentStatus?: AppointmentStatus;
  appointmentType?: AppointmentType;
  businessAffiliate?: BusinessAffiliateSummaryDto;
  createdAt?: string;
  id?: string;
  request?: RequestSummaryDto;
  updatedAt?: string;
  user?: UserDto;
}
// ...
export interface AddressDto {
  /** @nullable */
  city?: string | null;
  id?: string;
  /** @nullable */
  name?: string | null;
  state?: StatesAndTerritoriesOfMalaysia;
  /** @nullable */
  street?: string | null;
  /** @nullable */
  street2?: string | null;
  user?: UserDto;
  /** @nullable */
  zipCode?: string | null;
}